add IndirectMessage->all_parameters
removed
   # NOTE: There is intentionally no way to get all of the keys in the core
   # namespace because that means we don't need to be able to enumerate
   # to support the core protocol, and there is no requirement to enumerate
   # anyway.

because, as of OpenID 2.0, this is now actively wrong.

A 2.0 check_authentication message needs to include ALL parameters that
were received in the positive assertion response received and the only
way to do that is to be able to enumerate all of the keys.

Fortunately, in the one case where we couldn't do this before
(passing in a coderef as the URL parameter object), it's easy to
modify the contract (i.e., the new requirement that a coderef call
with no arguments returns the full list of URL parameters) -- and also
easy to recognize when there's legacy code not following it (parameter
names are never going to be undef, nor will there *ever* be zero
parameters since we always have .mode).